# .swagger/swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/images/upload":{"post":{"operationId":"ImageController_uploadImages","parameters":[],"responses":{"201":{"description":""}},"tags":["images"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthenticationController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCredentials"}}}},"responses":{"201":{"description":""}},"tags":["authentication"]}},"/api/v1/settings":{"get":{"operationId":"SettingsController_get","parameters":[],"responses":{"200":{"description":"successful get global dress options","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GlobalDressOptionsDto"}}}}},"tags":["settings"]},"post":{"operationId":"SettingsController_save","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GlobalDressOptionsDto"}}}},"responses":{"200":{"description":"successful save global dress options","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GlobalDressOptionsDto"}}}}},"tags":["settings"]}},"/api/v1/products/{id}":{"get":{"operationId":"ProductsController_getProduct","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"successful get product response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductViewDto"}}}}},"tags":["products"]},"delete":{"operationId":"ProductsController_deleteProduct","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["products"]}},"/api/v1/products/create":{"post":{"operationId":"ProductsController_createProduct","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}},"responses":{"200":{"description":"successful create product response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductViewDto"}}}}},"tags":["products"]}},"/api/v1/products/update/{id}":{"put":{"operationId":"ProductsController_updateProduct","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}},"responses":{"200":{"description":"successful update product response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductViewDto"}}}}},"tags":["products"]}},"/api/v1/products/search":{"post":{"operationId":"ProductsController_search","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductSearchDto"}}}},"responses":{"200":{"description":"successful search product response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductSearchViewDto"}}}}},"tags":["products"]}},"/api/v1/blog/{id}":{"get":{"operationId":"BlogController_getBlog","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"successful get blog response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogDto"}}}}},"tags":["blog"]}},"/api/v1/blog/search":{"post":{"operationId":"BlogController_searchBlog","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogSearchDto"}}}},"responses":{"200":{"description":"successful search blog response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogDto"}}}}},"tags":["blog"]}},"/api/v1/blog/uploadFile":{"post":{"operationId":"BlogController_uploadBlogFile","parameters":[],"responses":{"200":{"description":"successful upload blog file response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogUploadResultDto"}}}}},"tags":["blog"]}},"/api/v1/blog/create":{"post":{"operationId":"BlogController_createBlog","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogDto"}}}},"responses":{"200":{"description":"successful create blog response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogDto"}}}}},"tags":["blog"]}},"/api/v1/blog/update/{id}":{"put":{"operationId":"BlogController_updateBlog","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogDto"}}}},"responses":{"200":{"description":"successful update blog response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogDto"}}}}},"tags":["blog"]}},"/api/v1/blog/delete/{id}":{"delete":{"operationId":"BlogController_deleteBlog","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"successful delete blog response"}},"tags":["blog"]}},"/api/v1/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}}},"tags":["health"]}}},"info":{"title":"happy-dress-server","description":"happy-dress-server API description","version":"1.0","contact":{}},"tags":[{"name":"authentication","description":""},{"name":"health","description":""},{"name":"images","description":""},{"name":"products","description":""},{"name":"settings","description":""}],"servers":[],"components":{"schemas":{"UserCredentials":{"type":"object","properties":{"login":{"type":"string"},"password":{"type":"string"}},"required":["login","password"]},"CategoryDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"type":"string"},"imageUrl":{"type":"string"}},"required":["name","description","imageUrl"]},"ModelDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"}},"required":["name"]},"MaterialDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"}},"required":["name"]},"ColorDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"firstColor":{"type":"string"},"secondColor":{"type":"string"}},"required":["name","firstColor"]},"SizeDto":{"type":"object","properties":{"id":{"type":"number"},"sizeValue":{"type":"number"}},"required":["sizeValue"]},"GlobalDressOptionsDto":{"type":"object","properties":{"categories":{"type":"array","items":{"$ref":"#/components/schemas/CategoryDto"}},"models":{"type":"array","items":{"$ref":"#/components/schemas/ModelDto"}},"materials":{"type":"array","items":{"$ref":"#/components/schemas/MaterialDto"}},"colors":{"type":"array","items":{"$ref":"#/components/schemas/ColorDto"}},"sizes":{"type":"array","items":{"$ref":"#/components/schemas/SizeDto"}}},"required":["categories","models","materials","colors","sizes"]},"ProductColorSizeViewDto":{"type":"object","properties":{"color":{"$ref":"#/components/schemas/ColorDto"},"size":{"$ref":"#/components/schemas/SizeDto"}},"required":["color","size"]},"ProductColorImageViewDto":{"type":"object","properties":{"color":{"$ref":"#/components/schemas/ColorDto"},"imageURLs":{"type":"array","items":{"type":"string"}}},"required":["color","imageURLs"]},"ProductViewDto":{"type":"object","properties":{"id":{"type":"number"},"description":{"type":"string"},"mainImageUrl":{"type":"string"},"category":{"$ref":"#/components/schemas/CategoryDto"},"model":{"$ref":"#/components/schemas/ModelDto"},"materials":{"type":"array","items":{"$ref":"#/components/schemas/MaterialDto"}},"productColorSizes":{"type":"array","items":{"$ref":"#/components/schemas/ProductColorSizeViewDto"}},"productColorImages":{"type":"array","items":{"$ref":"#/components/schemas/ProductColorImageViewDto"}},"name":{"type":"string"}},"required":["description","mainImageUrl","category","model","materials","productColorSizes","productColorImages","name"]},"ProductColorImageDto":{"type":"object","properties":{"colorId":{"type":"number"},"imageURLs":{"type":"array","items":{"type":"string"}}},"required":["colorId","imageURLs"]},"ProductDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"type":"string"},"mainImageUrl":{"type":"string"},"categoryId":{"type":"number"},"modelId":{"type":"number"},"materialIds":{"type":"array","items":{"type":"number"}},"productColorImages":{"type":"array","items":{"$ref":"#/components/schemas/ProductColorImageDto"}},"productColorSizes":{"type":"array","items":{"$ref":"#/components/schemas/ProductColorImageDto"}}},"required":["name","description","mainImageUrl","categoryId","modelId","materialIds","productColorImages","productColorSizes"]},"ProductSearchDto":{"type":"object","properties":{"limit":{"type":"number"},"page":{"type":"number"},"categoryId":{"type":"number"},"modelIds":{"type":"array","items":{"type":"number"}},"materialIds":{"type":"array","items":{"type":"number"}},"colorIds":{"type":"array","items":{"type":"number"}},"sizeIds":{"type":"array","items":{"type":"number"}},"name":{"type":"string"}},"required":["limit","page"]},"ProductSearchViewDto":{"type":"object","properties":{"products":{"type":"array","items":{"$ref":"#/components/schemas/ProductViewDto"}},"totalItems":{"type":"number"},"itemsPerPage":{"type":"number"},"totalPages":{"type":"number"},"currentPage":{"type":"number"}},"required":["products","totalItems","itemsPerPage","totalPages","currentPage"]},"BlogDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"shortDescription":{"type":"string"},"isPublished":{"type":"boolean"},"htmlLinkBlog":{"type":"string"}},"required":["name","shortDescription","isPublished","htmlLinkBlog"]},"BlogSearchDto":{"type":"object","properties":{"name":{"type":"string"},"shortDescription":{"type":"string"}}},"BlogUploadResultDto":{"type":"object","properties":{}}}}}